Split out buildExec
